Add more theming e2e test cases
